Move the Monitoring folder inside this app
This is not really ideal, but since we need rather urgently to
run on Windows which doesn't have symlinks, and also package a
complete solution for quick installation, this is the stopgap
for now.

Longer term, need to let the user configure better where the
storage should live (e.g. if they want it to be on an external
or secondary drive)