Correctly (?!?) detect local IP and display it
On startup, the command line version will now set sensible
defaults for all available configuration variables; the default
station is "DEMO" and the default shared secret password is
"demo." The IP address in use, and the URL to visit to view
maps, is also printed out.

The auto-set, auto-detected values can still be overridden by
environment variables -- the names of the appropriate environment
variables are printed out on the console with these messages.