Support for syncing data to Community Lands.
Submission data can now be uploaded to a Community Lands
server.  In order to enable this, the following variables
must be provided in .env (or equivalent):

* community_lands_token - authentication token from CL website
* community_lands_server - the host name of CL (will be optional)
* community_lands_port - host port (optional)

Currently, the first two are required.

The workflow as of this commit is as follows:

* A request is made to the CL server to get the timestamp for the
  last time a sync was done
* Files are gathered, filtered based on said timestamp
* Files are sent over to the server for sync.

Also exposed is a means of re-syncing all files, ignoring the
timestamp.