Offline cache support for Bing satellite tile layer
This includes a modified version of Mapfilter that routes all
Bing requests -- both metadata and tiles -- through this service.

If the system is online, the requests are passed through to Bing
and serviced by Bing, with all tiles also being cached locally.
Locally cached tiles are always served locally regardless of
online status.

Metadata requests (needed to bootstrap the Bing layer) are
forwarded to Bing whenever the system is online. If this fails,
an old metadata is served up from the software image. This at
least allows the layer to initialize and cached tiles to serve.