Update settings page to new interface style.
Changed the h6's to h4's, removed the gray coloring, and
shifted the sections into their on wells, such that the
interface is consistent with the other tabs.

Also, added clearfix to the bottom of the page to compensate
for this change and keep the save button from the bottom of
the page.

Finally, removed the Application Settings header, as this now
seems superfluous. Left commented if needed back.