Report community lands upload progress.
Added a ServerEvent module responsible for transmitting
event information from the server to the main process. Then,
added a check on the Community Lands backup call to fire
events regarding the status of the upload. This information
is listened for and displayed to the client.

This information is regarding the upload progress ONLY; that
is to say, once the file is uploaded to Community Lands, no
further updates will be given until Community Lands has
processed the zip file and the request has completed.