Externalize formerly inline index javascript
This should not provoke any regressions and is meant to be
syntactically identical to the previous version -- just instead
of one big <script>...</script> block in the index.html file,
this externalizes all the same into an actions.js file that
is still read into global namespace.

I ran the externalized file through standard-js and made it
jslint warning clean, but other than that cosmetic cleanup this
should not have altered its functionality at all.

Modularizing the loading and status screens so they don't
pollute global namespace, and modularizing this so it doesn't
either, would be logical next steps.