Move the CMS features out to a module
This includes changes to loading and locale to *allow* them to be
required as a module, although they also continue to be present
in the global scope for compatibility with old code.

cms.js uses the UJS style where the code itself handles the
binding back to the UI, as opposed to the UI being able to invoke
the code.